Project ETL Pipeline/requirements.txt
absl-py==2.2.2
ace_tools==0.0
altair==5.5.0
astroid==3.3.8
asttokens==3.0.0
attrs==24.2.0
autopep8==2.3.2
babel==2.17.0
beautifulsoup4==4.13.3
black==25.1.0
blinker==1.9.0
bqplot==0.12.43
bs4==0.0.2
cachetools==5.5.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
coverage==7.8.0
cycler==0.12.1
debugpy==1.8.9
decorator==5.1.1
dill==0.3.9
executing==2.1.0
fastjsonschema==2.21.1
flake8==7.1.2
Flask==3.1.0
fonttools==4.55.2
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.24.2
google-api-python-client==2.169.0
google-auth==2.40.1
google-auth-httplib2==0.2.0
google-play-scraper==1.2.7
googleapis-common-protos==1.70.0
greenlet==3.2.1
h11==0.16.0
h5py==3.13.0
httplib2==0.22.0
idna==3.10
iniconfig==2.1.0
ipydatawidgets==4.3.5
ipykernel==6.29.5
ipython==8.30.0
ipython-genutils==0.2.0
ipyvolume==0.6.3
ipyvue==1.11.2
ipyvuetify==1.10.0
ipywebrtc==0.6.0
ipywidgets==8.1.5
isort==6.0.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
keras==3.9.2
kiwisolver==1.4.7
lxml==5.3.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.3
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
ml_dtypes==0.5.1
mypy-extensions==1.0.0
namex==0.0.8
narwhals==1.29.0
nbformat==5.10.4
nest-asyncio==1.6.0
nltk==3.9.1
numpy==2.1.3
optree==0.15.0
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pathspec==0.12.1
patsy==1.0.1
pillow==11.0.0
platformdirs==4.3.6
plotly==5.24.1
pluggy==1.5.0
prompt_toolkit==3.0.48
proto-plus==1.26.1
protobuf==5.29.3
psutil==6.1.0
pure_eval==0.2.3
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.12.1
pycparser==2.22
pydeck==0.9.1
pyflakes==3.2.0
Pygments==2.18.0
pylint==3.3.4
pyparsing==3.2.0
PySocks==1.7.1
pytest==8.3.5
pytest-cov==6.1.1
python-dateutil==2.9.0.post0
pythreejs==2.4.2
pytz==2024.2
pywin32==308
pyzmq==26.2.0
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rich==14.0.0
rpds-py==0.22.3
rsa==4.9.1
Sastrawi==1.0.1
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
selenium==4.32.0
setuptools==78.0.1
simpy==4.1.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.40
stack-data==0.6.3
statsmodels==0.14.4
streamlit==1.43.0
tenacity==9.0.0
textblob==0.18.0.post0
threadpoolctl==3.5.0
toml==0.10.2
tomlkit==0.13.2
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
traittypes==0.2.1
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
watchdog==6.0.0
wcwidth==0.2.13
websocket-client==1.8.0
Werkzeug==3.1.3
widgetsnbextension==4.0.13
wordcloud==1.9.4
wsproto==1.2.0
yapf==0.43.0
yellowbrick==1.5
